+++ apps/controllo-ore-x/src/app/_core/services/index-configuration.data-service.ts
